typecasting output of procs to floats to match the schema

+++ procs/stats.py
@@ -49,6 +49,6 @@ def proc_main(pk, row, arg):
 
     """
     print("[[[[[[[[[[[STATS]]]]]]]]]]]]", pk, row, arg)
-    damean = row['ts'].mean()
-    dastd = row['ts'].std()
+    damean = float(row['ts'].mean())
+    dastd = float(row['ts'].std())
     return [damean, dastd]
